Will the member attend the fitness class?
PURPOSE OF PREPARING THE MODEL:
GoalZone is a fitness club chain in Canada. It offers a range of fitness classes in two capacities - 25 and 15. Some classes are always fully booked. Fully booked classes often have a low attendance rate. GoalZone wants to increase the number of spaces available for classes. They want to do this by predicting whether the member will attend the class or not. If they can predict that a member will not attend the class, they can make another space available. This model is prepared using two algorithms: Logistic Regression & Decision Tree Classifier.